# SRS caching