--- a/src/index.ts
+++ b/src/index.ts
@@ -45,7 +45,7 @@ export {
   Undefined,
   Void,
 } from './lib/proof_system.js';
-export { Cache } from './lib/proof-system/cache.js';
+export { Cache, CacheHeader } from './lib/proof-system/cache.js';
 
 export {
   Token,