@mitschabaude 2023-10-31
Damn, it fails in the web version, I forgot to consider the weird worker-calling layer we have there 😞 Might need to refactor this a bit to work